// src/frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface Word {
    id: bigint;
    difficulty: Difficulty;
    languageName: string;
    addedAt: Time;
    foreign: string;
    examples: Array<string>;
    english: string;
}
// ...
export enum Difficulty {
    beginner = "beginner",
    advanced = "advanced",
    hard = "hard",
    medium = "medium"
}